+++ client/web/BUILD.bazel
@@ -204,6 +204,11 @@ ts_project(
         "src/cody/chat/CodyChatPage.tsx",
         "src/cody/chat/CodyPageIcon.tsx",
         "src/cody/chat/index.tsx",
+        "src/cody/chat/new-chat/NewCodyChatPage.tsx",
+        "src/cody/chat/new-chat/components/chat-history-list/ChatHistoryList.tsx",
+        "src/cody/chat/new-chat/components/chat-ui/ChatUi.tsx",
+        "src/cody/chat/new-chat/components/skeleton/Skeleton.tsx",
+        "src/cody/chat/old-chat/CodyChatPage.tsx",
         "src/cody/codyProRoutes.tsx",
         "src/cody/components/ChatUI/ChatUi.tsx",
         "src/cody/components/ChatUI/index.tsx",
@@ -265,6 +270,9 @@ ts_project(
         "src/cody/sidebar/CodySidebar.tsx",
         "src/cody/sidebar/Provider.tsx",
         "src/cody/sidebar/index.tsx",
+        "src/cody/sidebar/new-cody-sidebar/NewCodySidebar.tsx",
+        "src/cody/sidebar/new-cody-sidebar/NewCodySidebarWebChat.tsx",
+        "src/cody/sidebar/old-cody-sidebar/CodySidebar.tsx",
         "src/cody/sidebar/useSidebarSize.tsx",
         "src/cody/subscription/CodySubscriptionPage.tsx",
         "src/cody/subscription/queries.tsx",
@@ -1824,6 +1832,7 @@ ts_project(
         "//:node_modules/apollo3-cache-persist",
         "//:node_modules/bloomfilter",
         "//:node_modules/classnames",
+        "//:node_modules/cody-web-experimental",
         "//:node_modules/copy-to-clipboard",
         "//:node_modules/d3-scale-chromatic",
         "//:node_modules/d3-time-format",
