+++ client/web/dev/esbuild/config.ts
@@ -48,6 +48,12 @@ export function esbuildBuildOptions(ENVIRONMENT_CONFIG: EnvironmentConfig): esbu
             workerPlugin,
             packageResolutionPlugin({
                 path: require.resolve('path-browserify'),
+                // TODO(sqs): force use of same version when developing on opencodegraph because `pnpm link` breaks
+                '@codemirror/state': path.join(ROOT_PATH, 'node_modules/@codemirror/state'),
+                '@codemirror/view': path.join(ROOT_PATH, 'node_modules/@codemirror/view'),
+                react: path.join(ROOT_PATH, 'node_modules/react'),
+                'react-dom': path.join(ROOT_PATH, 'node_modules/react-dom'),
+                'react-dom/client': path.join(ROOT_PATH, 'node_modules/react-dom/client'),
                 ...(ENVIRONMENT_CONFIG.DEV_WEB_BUILDER_OMIT_SLOW_DEPS
                     ? {
                           // Monaco
@@ -64,13 +70,6 @@ export function esbuildBuildOptions(ENVIRONMENT_CONFIG: EnvironmentConfig): esbu
 
                           // Misc.
                           recharts: '/dev/null',
-
-                          // TODO(sqs): force use of same version when developing on opencodegraph because `pnpm link` breaks
-                          '@codemirror/state': path.join(ROOT_PATH, 'node_modules/@codemirror/state'),
-                          '@codemirror/view': path.join(ROOT_PATH, 'node_modules/@codemirror/view'),
-                          react: path.join(ROOT_PATH, 'node_modules/react'),
-                          'react-dom': path.join(ROOT_PATH, 'node_modules/react-dom'),
-                          'react-dom/client': path.join(ROOT_PATH, 'node_modules/react-dom/client'),
                       }
                     : null),
             }),
