+++ client/web/src/cody/chat/new-chat/components/skeleton/Skeleton.tsx
@@ -1,0 +1,14 @@
+import type { FC } from 'react'
+
+import classNames from 'classnames'
+
+import styles from './Skeleton.module.scss'
+
+export const Skeleton: FC = props => (
+    <div className={styles.lines}>
+        <div className={classNames(styles.line, styles.lineTitle)} />
+        <div className={styles.line} />
+        <div className={styles.line} />
+        <div className={styles.line} />
+    </div>
+)
