+++ client/web/src/cody/chat/old-chat/CodyChatPage.module.scss
