+++ client/web/src/cody/sidebar/old-cody-sidebar/CodySidebar.module.scss
