+++ client/web/src/repo/RepoContainer.module.scss
@@ -3,3 +3,7 @@
     margin-right: auto;
     overflow-x: hidden;
 }
+
+.cody-sidebar-panel {
+    border-left: 1px solid var(--border-color);
+}
